Simplify gexiv2 API: Closes #6170
This cleans up the API in various ways, as described in the above
ticket.  All functionality is still available, it's just that some
of the work is pushed out to the client user (basically, selecting
the Exiv2 tag they want to work with) in the interests of keeping
the interface clean.

This interface change requires no changes to Shotwell, but may affect
other gexiv2 clients.