Port gexiv2 to autotools: Closes #7660
This replaces gexiv2's old homebrew Makefile with an autotools
equivalent that matches, almost option-for-option, the old available
configuration options.  It also moves one of the filenames to a more
standard convention (gexiv2.m4 -> gexiv2.pc.in).

The VAPI is also updated here to include the new version number API
(#7647, commit:b6496c76).