Fix GPS rational calculation
Apparently there are some images where num == den == 0 for the GPS fractions
which caused the code to bail out too early and left a) wrong GPS values and
b) caused get_gps_info to return 0 for all other values after the failing one.

While we're at it, clean up the functions to be a bit more C++-y. Just because
we export a C interface doesn't mean we can't use nice C++ internally.

Signed-off-by: Jens Georg <mail@jensge.org>

https://bugzilla.gnome.org/show_bug.cgi?id=775249