gexiv2: more variants with GError instead of WARNINGs.
This gets rid of all functions which could fail with a LOG_ERROR(),
which was basically raising a GLib warning. Use GError instead as this
is much easier to process for using applications.

For API stability, it results in new public functions, while the
existing ones keep behaving the same yet are now deprecated in favor of
the new ones.