tools: fix gexiv2-dump build.
metadata.get_tag_interpreted_string() outputs a warning on the CI
because it is deprecated.
And metadata.try_get_tag_interpreted_string() outputs another warning
because (at the opposite) is it "not available in gexiv2 0.12.1. Use
gexiv2 >= 0.12.2" (which is normal as we are a dev version leading to
0.12.2).
So adding --disable-since-check to vala_args (which I found by testing
randomly as vala_args parameter to executable() is documented nowhere on
meson docs!).